Move MultiFacilityImpactDashboard into its own component
