+++ src/page-multi-facility/MultiFacilityImpactDashboard.tsx
@@ -1,0 +1,63 @@
+import React, { useEffect, useState } from "react";
+import styled from "styled-components";
+
+import { getFacilities } from "../database";
+import Loading from "../design-system/Loading";
+import AddFacilityModal from "./AddFacilityModal";
+import { FacilityContext } from "./FacilityContext";
+import FacilityRow from "./FacilityRow";
+import ProjectionsHeader from "./ProjectionsHeader";
+import ScenarioSidebar from "./ScenarioSidebar";
+import { Facilities } from "./types";
+
+const MultiFacilityImpactDashboardContainer = styled.main.attrs({
+  className: `
+    h-screen
+    flex
+    mt-8
+  `,
+})``;
+
+const MultiFacilityImpactDashboard: React.FC = () => {
+  const [facilities, setFacilities] = useState({
+    data: [] as Facilities,
+    loading: true,
+  });
+
+  useEffect(() => {
+    async function fetchFacilities() {
+      const facilitiesData = await getFacilities();
+
+      if (facilitiesData) {
+        setFacilities({
+          data: facilitiesData,
+          loading: false,
+        });
+      }
+    }
+    fetchFacilities();
+  }, []);
+
+  return (
+    <MultiFacilityImpactDashboardContainer>
+      <ScenarioSidebar />
+      <div className="flex flex-col flex-1 pb-6 pl-8">
+        <AddFacilityModal />
+        <ProjectionsHeader />
+        {facilities.loading ? (
+          <Loading />
+        ) : (
+          facilities?.data.map((facility, index) => {
+            return (
+              <FacilityContext.Provider key={index} value={facility}>
+                <FacilityRow />
+              </FacilityContext.Provider>
+            );
+          })
+        )}
+      </div>
+    </MultiFacilityImpactDashboardContainer>
+  );
+};
+
+export default MultiFacilityImpactDashboard;
